# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Travel Agent API
  description: API for conversational flight search with price monitoring and alerts
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.travelagent.example.com/v1
    description: Production server
...
paths:
...
  /flights/search:
    post:
      tags:
        - Flights
      summary: Search for flights
      operationId: searchFlights
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlightSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightSearchResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /flights/saved-searches:
    get:
      tags:
        - Flights
      summary: Get saved searches
      operationId: getSavedSearches
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of saved searches
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavedSearch'
    
    post:
      tags:
        - Flights
      summary: Save a search
      operationId: saveSearch
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSearchRequest'
      responses:
        '201':
          description: Search saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      message:
                        type: string
  /flights/saved-searches/{id}:
    delete:
      tags:
        - Flights
      summary: Delete saved search
      operationId: deleteSavedSearch
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search deleted
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
  schemas:
...
    FlightSearchRequest:
      type: object
      required:
        - origin
        - destination
        - departureDate
      properties:
        origin:
          type: string
          pattern: '^[A-Z]{3}$'
        destination:
          type: string
          pattern: '^[A-Z]{3}$'
        departureDate:
          type: string
          format: date
        returnDate:
          type: string
          format: date
        adults:
          type: integer
          minimum: 1
          maximum: 9
          default: 1
        children:
          type: integer
          minimum: 0
          maximum: 8
          default: 0
        infants:
          type: integer
          minimum: 0
          maximum: 8
          default: 0
        cabinClass:
          type: string
          enum: [ECONOMY, PREMIUM_ECONOMY, BUSINESS, FIRST]
          default: ECONOMY
        nonStop:
          type: boolean
          default: false
        maxPrice:
          type: number
        currency:
          type: string
          default: USD
    FlightSearchResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            offers:
              type: array
              items:
                $ref: '#/components/schemas/FlightOffer'
            meta:
              type: object
              properties:
                count:
                  type: integer
    FlightOffer:
      type: object
      properties:
        id:
          type: string
        price:
          type: object
          properties:
            currency:
              type: string
            total:
              type: string
            base:
              type: string
            grandTotal:
              type: string
        itineraries:
          type: array
          items:
            type: object
        validatingAirlineCodes:
          type: array
          items:
            type: string
    SavedSearch:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        searchParams:
          type: object
        priceAlerts:
          type: object
          properties:
            enabled:
              type: boolean
            targetPrice:
              type: number
            lastCheckedPrice:
              type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SaveSearchRequest:
      type: object
      required:
        - name
        - searchParams
      properties:
        name:
          type: string
        searchParams:
          type: object
        priceAlerts:
          type: object
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...